import-python/pyproject.toml
[tool.poetry]
name = "import-python"
version = "0.1.0"
description = ""
authors = ["Ben Gamari <ben@well-typed.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25"
asyncpg = "^0.22"
python-gitlab = "^2.6"
python-dateutil = "^2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
